Render a minesweeper game
This game was copied from
https://hackernoon.com/build-your-own-react-48edb8ed350d
It works in our React, but renders alot slower than the React from the
blog post. Also, just like with the React from the blog post, you can not
change the board size. Not sure if this is a problem with our React or
with the code in the game itself.